Added the 'amount' column to the transactions table
This improves the performance greatly when purchasing multiple
beverages. Before, 5 beverages of the same type would need 5 inserts,
now it's only 1 insert with an amount of 5.
This should also speed up calculating the prices as less database reads
need to happen.

+++ src/transactions/routes.rs
@@ -31,7 +31,7 @@ async fn get_sales(game_id: Path<i64>, id: Identity, pool: Data<db::Pool>) -> se
 #[post("/games/{id}/sales")]
 async fn create_sale(
     game_id: Path<i64>,
-    slots: Json<HashMap<i16, u8>>,
+    slots: Json<HashMap<i16, i32>>,
     id: Identity,
     pool: Data<db::Pool>,
     tx: Data<mpsc::Sender<Sale>>,
